+++ src/main/java/kitchenpos/menugroup/domain/MenuGroup.java
@@ -1,6 +1,5 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menugroup.domain;
 
-import java.math.BigDecimal;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -26,13 +25,6 @@ public class MenuGroup {
     public MenuGroup(final Long id, final String name) {
         this.id = id;
         this.name = name;
-    }
-
-    public Menu createMenu(final String name, final BigDecimal price, final PendingMenuProducts products) {
-        if (price == null || price.compareTo(BigDecimal.ZERO) < 0 || price.compareTo(products.getTotalPrice()) > 0) {
-            throw new IllegalArgumentException();
-        }
-        return new Menu(name, price, id, products.createMenuProducts());
     }
 
     public Long getId() {
