+++ src/main/java/kitchenpos/menugroup/domain/MenuGroupRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.menu.domain.repository;
+package kitchenpos.menugroup.domain;
 
-import kitchenpos.menu.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
