+++ src/test/java/acceptance/MenuGroupAcceptanceTest.java
@@ -12,8 +12,8 @@ import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import common.AcceptanceTest;
 import io.restassured.RestAssured;
 import java.util.List;
-import kitchenpos.menu.ui.request.MenuGroupRequest;
-import kitchenpos.menu.ui.response.MenuGroupResponse;
+import kitchenpos.menugroup.ui.request.MenuGroupRequest;
+import kitchenpos.menugroup.ui.response.MenuGroupResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
